# Code clarity (path strings) and enhanced performance