--- a/salted/aims/make_geoms.py
+++ b/salted/aims/make_geoms.py
@@ -1,15 +1,18 @@
-import inp
 import os
-from ase.io import read, write
 import argparse
+import os.path as osp
 
-def add_command_line_arguments_contraction(parsetext):
-    parser = argparse.ArgumentParser(description=parsetext)
+from ase.io import read, write
+
+import inp
+
+def add_command_line_arguments_contraction():
+    parser = argparse.ArgumentParser()
     parser.add_argument("-pr", "--predict", action='store_true', help="Prepare geometries for a true prediction")
     args = parser.parse_args()
     return args
 
-args = add_command_line_arguments_contraction("")
+args = add_command_line_arguments_contraction()
 predict = args.predict
 
 fname = inp.filename
@@ -18,15 +21,11 @@ if predict:
 else:
     datadir = "data/"
 
-if not os.path.exists(inp.path2qm):
-    os.mkdir(inp.path2qm)
-if not os.path.exists(inp.path2qm+datadir):
-    os.mkdir(inp.path2qm+datadir)
-dirpath = inp.path2qm+datadir+'geoms/'
-if not os.path.exists(dirpath):
-    os.mkdir(dirpath)
+dirpath = osp.join(inp.path2qm, datadir, "geoms")
+if not osp.exists(dirpath):
+    os.makedirs(dirpath)
 
 xyz_file = read(fname,":")
 n = len(xyz_file)
 for i in range(n):
-    write(dirpath+str(i+1)+'.in',xyz_file[i])
+    write(osp.join(dirpath, f"{i+1}.in"), xyz_file[i])