--- a/salted/aims/move_data_in.py
+++ b/salted/aims/move_data_in.py
@@ -9,18 +9,23 @@ from salted.sys_utils import read_system
 
 species, lmax, nmax, lmax_max, nnmax, ndata, atomic_symbols, natoms, natmax = read_system()
 
-pdir = 'predictions_'+inp.saltedname+'_'+inp.predname+'/'
+pdir = f"predictions_{inp.saltedname}_{inp.predname}"
 
 M = inp.Menv
 ntrain = int(inp.trainfrac*inp.Ntrain)
 
 for i in range(ndata):
-    t = np.load(inp.saltedpath+pdir+"M"+str(M)+"_zeta"+str(inp.z)+"/N"+str(ntrain)+"_reg"+str(int(np.log10(inp.regul)))+"/prediction_conf"+str(i)+".npy")
+    print(f"processing {i+1}/{ndata} frame")
+    t = np.load(os.path.join(
+        inp.saltedpath, pdir,
+        f"M{M}_zeta{inp.z}", f"N{ntrain}_reg{int(np.log10(inp.regul))}",
+        f"prediction_conf{i}.npy",
+    ))
     n = len(t)
 
-    dirpath = os.path.join(inp.path2qm, inp.predict_data, str(i+1))
+    dirpath = os.path.join(inp.path2qm, inp.predict_data, f"{i+1}")
 
-    idx = np.loadtxt(dirpath+'/idx_prodbas.out').astype(int)
+    idx = np.loadtxt(os.path.join(dirpath, f"idx_prodbas.out")).astype(int)
     idx -= 1
     idx = list(idx)
     idx_rev = []
@@ -28,7 +33,9 @@ for i in range(ndata):
         idx_rev.append(idx.index(i))
     #	np.savetxt('../idx_prodbas_rev.out',idx_rev)
     #	idx_rev = np.loadtxt('../idx_prodbas_rev.out').astype(int)
-    cs_list = np.loadtxt(dirpath+'/prodbas_condon_shotley_list.out').astype(int)
+    cs_list = np.loadtxt(os.path.join(
+        dirpath, f"prodbas_condon_shotley_list.out"
+    )).astype(int)
     cs_list -= 1
     cs_list = list(cs_list)
 
@@ -36,5 +43,5 @@ for i in range(ndata):
     for j in cs_list:
         t[j] *= -1
 
-    np.savetxt(dirpath+'/ri_restart_coeffs_predicted.out',t)
+    np.savetxt(os.path.join(dirpath, f"ri_restart_coeffs_predicted.out"), t)