--- a/salted/minimize_loss.py
+++ b/salted/minimize_loss.py
@@ -51,13 +51,10 @@ def build():
         for spe in species:
             av_coefs[spe] = np.load(f"averages_{spe}.npy")
 
+    dirpath = os.path.join(inp.saltedpath, rdir, f"M{M}_zeta{zeta}")
     if rank==0:
-        dirpath = os.path.join(inp.saltedpath, rdir)
         if not os.path.exists(dirpath):
-            os.mkdir(dirpath)
-        dirpath = os.path.join(inp.saltedpath, rdir, f"M{M}_zeta{zeta}")
-        if not os.path.exists(dirpath):
-            os.mkdir(dirpath)
+            os.makedirs(dirpath)
     if size > 1: comm.Barrier()
 
     # define training set at random