Dear development team,

This pull request focuses on improving clarity of path strings (as what Andrea @andreagrisafi  agrees with), and some performance optimization (I mentioned these to Mariana). Key updates include:

1. **Path string concatenation**: Utilization of `os.path.join()` and `os.makedirs()` has been applied for a more robust and easy-understanding path string concatenation and subdir creating.
1. **Performance enhancements**: Significant improvements have been made in two critical files:
    - **File**: `salted/aims/get_df_err.py`
        - **Improvement**: Implementation of a more efficient method for spatial grid sorting.
        - **Benchmark Results**:
            - Time reduction from 94.7ms to 21.7ms in a specific code segment attached below.
            - On the first 100 training samples of the ZrS2 dataset (Zr2S4), total execution time decreased from 527s to 408s without impacting the accuracy of results.
    - **File**: `salted/aims/move_data_in.py`
        - **Enhancements**: Optimizations in index reversing and parity reversing segments.
        - **Benchmark Results**:
            - On ZrS2 prediction dataset frame 1 (Zr38S76), the predicted density fitting coefficients array length is 46969
            - Index Reversing: Time reduced dramatically from 113s to 14ms.
            - Parity Reversing: Time decreased from 19ms to 3ms.
            - A comprehensive timing comparison is included for reference (see below).
1. **Code Refinement**: Minor f-string modifications have been introduced for cleaner code and improved terminal output. These changes are not expected to affect computational results.

I didn't change the cp2k- and pyscf-only code. I have verified this PR by running SALTED with FHI-aims on the ZrS2 dataset.

> **24.01.24 update**: (improving the path strings) I have revised all files in the directory `salted`, except for subdirs `cp2k` and `ortho`. These modifications start from commit `b07bb1b`.

I am confident these enhancements will contribute significantly to our project's efficiency and maintainability. I eagerly await your feedback and am available for any further discussion or clarification.
Thank you for your time and consideration.

Best regards,
Zekun

---

- spatial grid sorting
```python
# r_con is from ZrS2 dataset training frame 1

# before
%%timeit
r_con1 = r_con.copy()  # method in code now
r_con1.view('f8,f8,f8,f8').sort(order=['f0','f1','f2'],axis = 0)
>>> 94.7 ms ± 15.6 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# after
%%timeit
r_con2 = r_con.copy()  # improved method
r_con2 = r_con2[numpy.lexsort((r_con2[:, 2], r_con2[:, 1], r_con2[:, 0]))]
>>> 21.7 ms ± 79 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# reconfirm results
numpy.allclose(r_con1, r_con2, rtol=1e-8)
>>> True
```

- Index reversing and parity reversing
```python
# before, in seconds
time_load_pred=0.0007796287536621094
time_load_idx=0.009583234786987305
time_rev_idx=111.5866174697876  # <-- index reversing
time_load_cs=0.004029512405395508
time_rev_cs=0.018198728561401367  # <-- parity reversing
time_total=111.61933779716492
# after, in seconds
time_load_pred=0.0008151531219482422
time_load_idx=0.010459423065185547
time_rev_idx=112.59937787055969  # <-- index reversing
time_rev_idx_2=0.01427459716796875  # <-- optimized
test new method correct: True  # by numpy.allclose()
time_load_cs=0.0043964385986328125
time_rev_cs=0.018590450286865234  # <-- parity reversing
time_rev_cs_2=0.003393888473510742  # <-- optimized
test new method correct: True  # by numpy.allclose()
time_total=112.6575288772583
```